requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.2
Django==3.2.8
django-cors-headers==3.10.0
fpdf2==2.4.5
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
pytz==2021.3
regex==2021.10.8
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
